Handle quantified expressions correctly when building the graph.
Until now, quantifiers were not taken into consideration when creating
the graph. As a result, we could not calculate a minimal depth for a
recursive expression that was the only alternative of a rule -
even if it was guarded by a quantifier enabling its complete omission.
The patch fixes this.