Use minimum depths to guide quantified decisions.
Until now, the quantifier iterators didn't take the minimum depths
into consideration but only stopped the iteration if we already
exceeded the maximum depth. However, the correct behaviour would
have been to stop if the minimum depth of the quantifier's subtree
is larger than the currently available depth. The patch fixes this.