Clean up terminology around lexers, parsers, unlexers, and unparsers
Some terms were used confusingly or incorrectly: lexer was used for
unlexers, and parser was used for CLI parsers, ANTLR parser classes
and objects, and unparsers. The refactoring affects both tool code,
generated code, and runtime code.

As a side effect, some not-only-terminology-related code was also
changed in tool code.