Remove assert from node replace and change return type.
Since labelled alternatives are supported, node instances of the same
parser rule can have different names (since the name of the rule is
postfixed by the label that matches). In spite of this, the replace
mechanism still expected to replace nodes with exactly the same name.
The patch removes this expectation.
Besides, it also changes the return value from the replaced node to
the replacement node, since in most of the scenarios, the new node is
used later.