Merge unlexer and unparser concepts
Until now, we generated an unlexer and an unparser from the lexer and
parser grammars respectively, mirroring the behaviour of ANTLR's
parser generator mechanism. However, contrary to ANTLR's lexer and
parser classes, our unlexer and unparser don't have any semantic
differences and there is no dependency between them. This means that their
functionalities can be merged into a common class. The name of this
common class is postfixed by 'Generator'.
The patch also adapts the CLI of grammarinator-generate to use
this new single generator class, instead of the previously used
unlexer and unparser pairs.

Further changes:
  * changed name of the base class of generators (Grammarinator -> Generator)
  * removed max_cnt and node_cnt from Generator base class
  * moved the initialization of common fields from generated classes to Generator base class
  * white space changes