Stabilize order of functions in fuzzer generated by grammarinator-process
Rules of the grammar are represented by nodes of a graph, which
will be used to generate functions in the fuzzer. The nodes were
inserted in a dict, which has different behavior regarding order of
iteration accross different Python versions. (Up to py35, dict had
random iteration order. In py36, CPython had insertion order as an
implementation detail. Since py37, insertion order is part of the
language spec.) However, OrderedDict works consistently accross all
Python versions, so this patch makes the container of the nodes an
OrderedDict, thus stabilizing the order of the generated functions.