Rework the processing of charsets
Fixed inverted/negated/not-sets to use "dot" as the base charset
(always subtracted the specified charset from the set of printable
unicode characters previously, even if "dot" was the set of
printable ASCII characters). This required the move of the
computation of the charsets from runtime to processing-time. Which
enabled the simplification of the internal graph representation of
charsets and token/string literals. The codegen template was
adapted to the changed internal graph representation.

Also realized that Generator base class was still in charge of
randomly selecting characters from charsets. However, random
aspects (alternatives, quantifiers) are the responsibilities of
models. So, extended the models API with a method to deal with
charsets and moved the existing functionality from Generator to
DefaultModel. The codegen template was adapted to use the new API.