Add grammar analyzer functionality to grammarinator-process
This new funcionality helps identifying cycles in the graph and it
looks for unavailable rules from a user-defined start rule.
Furthermore, it finds the farthest rule from a starting rule
- this can be predefined or the first parser rule. This information
can be useful by generation when parametrizing the max-depth argument.
The node classes representing the grammar tree are extended with
__str__ and print_tree methods.