Remove ANTLRv4JavaLexer.g4 and LexerAdaptor.java
The generation of Java code from the ANTLRv4*.g4 grammar files was
never implemented, and it is not planned in the foreseeable future
either. (Java-based lexers/parsers have generally better
performance than their Python-based counterparts, but that's only
relevant if the input they are parsing is big and/or complex.
Grammar files in ANTLRv4 format usually don't fall in this
category.) Thus, this change removes the dead data from the
project.