Build-time generation of ANTLRv4Lexer and Parser
The ANTLRv4 lexer and parser are used to process .g4 grammar files
(from which fuzzers are created). However, the ANTLRv4 grammars,
and the lexer and parser generated from them are constant. Hence,
building the lexer and parser every time is redundant and a waste
of resources.

This patch moves the generation of the lexer and the parser from
run-time to build-time (using the build_antlr setuptools extension
provided by ANTLeRinator).