Prefix all hard-coded names in generators with an underscore
Prefixed all data members and methods of the base class Generator
as well as all hard-coded names in the generator template with an
underscore to avoid clashes with grammar rule names (which would
become method names in the instantiated template). This works as
no rule name can start with an underscore in ANTLR grammars.