Support args, locals and returns in parser rules
Similarly to ANTLRv4, Grammarinator supports the afforementioned
language elements for propagating information up (returns) and
down (args) in the tree during generation and declaring
variables in rule level to use them while generating it (locals).

Furthermore, the patch extends the handling of parser rule labels
with supporting the collection the labeled elements into a list
(`+=` operator with label definition).