Fix the mutation of inline tokens (#82)
Inline tokens are named as `<INVALID>` by ANTLR. When such a node
was chosen to be mutated then Grammarinator tried to regenerate
them according to a rule definition called `<INVALID>` which did
not exist, of course. The patch disables the mutation of such nodes
since inline tokens are constants, hence mutating them would not
have any importance either.

Fixes #80.