Add support for encoding error handling
The support for encoding error handling has been motivated by
errors caused if Unicode surrogate pairs are in the generated
tests, which cannot be written to output with utf-8 codec.

Also added a unit test to create a test case with character outside
the basic multilingual plane of Unicode, both in six-position-hex
format and with surrogate pair.