Better control of multiprocessing in unit tests
The default value for `-j` (number oj jobs) in `grammarinator-generate`
is based on the number of CPUs available on the system. This is
good (adaptive) in practice but not so good for testing (when
reproducibility is preferrable).

Therefore, `-j` is explicitly specified in all unit tests. All
unit tests try single-process test case generation (using `-j 1`)
and those test cases where multi-process behaviour needs to be
tested, do so (using `-j 2`).