Add listeners argument to Generator
This allows to set the list of listeners at generator construction
time.

The patch also retires the method add_listener, which had a naming
issue (i.e., any member in Generator that is public - does not
start with an underscore - may conflict later with a method of a
subclass named after a grammar rule).