Extend the depth-based size control with a token-based variant
Contrary to depth-based control, the token-based variant maintains
an accumulative limit value, which summarizes the token counts accross
the various levels of the tree. Hence, at every decision point we have
to know how many tokens are generated already. How many tokens are needed
at least to generate the current children and how many tokens are needed
on the rigth to finish the whole tree. Minimal token numbers are calculated
statically now, in the processor tool, while the generator dynamically keeps
track of the value of the current budget.