Fix the usage of rule.name and rule.id (#132)
The difference between the `name` and `id` properties of a rule is
that the `id` contains the value of a label if exists, too, which
results in a more accurate identifier for a rule. Hence, using the `id`
property instead of `name` for identifying the container rule of an
alternation, quantifier or a charset is reasonable. Similarly, using
it in the `__str__` method of a rule is more beneficial.