Improve the handling of the `dot` wildcard in parser rules (#134)
Until now, the `dot` wildcard was handled the same way both in lexer
and parser rules: it meant the generation of a single character.
But it was wrong. From now, if `dot` is placed in a parser rule,
it triggers the generation of a random token, as expected by ANTLR.
The only limitation is that lexer modes are not taken into account
(for now), since it would cause remarkable complication.