Pass around the in-edge of a node in the template
The existing `args` is less meaningful, as not all types of nodes
have "args", but they all have in-edges. Plus, should there be
additional data associated with the in-edge in the future, this
helps avoid the need for passing it around separately.

(NB: Even `node` is a field of the in-edge -- its `dst` -- but it
is used very often, so having it as `node` instead of `inedge.dst`
keeps referring to it simpler.)