Don't shadow builtin min and max in quantifier-related code (#163)
Both the tools (the processor, the codegen template) and the
runtime (the models) contain code that deals with quantifiers.
Quantification has lower and upper bounds, which are traditionally
named min and max. However, having min and max as function
arguments (like in models) shadows the Python builtins. To avoid
this and make the builtins available, min is renamed to start while
max is renamed to stop all around the code base (terminology
borrowed from the range Python builtin).