Add support for multiple tree file formats
The now-added tree codec framework provides means to convert
between trees and arrays of bytes. Moreover, enhanced codecs can
also support the serialization and deserialization of additional
data (called annotations). The default population implementation
is updated to make use of the tree codec framework and make optimal
use of the annotations if supported by the chosen codec.

At the moment, two tree codecs are supported: a json-based, with no
support for annotations, and a pickle-based, with support for
annotations.

Note: The internal format of the files read and written by the
pickle-based tree codec are NOT backward compatible with the
serialized trees created with previous versions.

Co-authored-by: Renata Hodovan <reni@inf.u-szeged.hu>