Enable labeled alternatives to start their name with their rule name
In ANTLR, there is no such restriction on the name of a labeled
alternative that it cannot start with the name of its container
rule. The only restriction is that it cannot match completly with
any rule name. The patch removes such a mistaken restriction
from the Grammarinator parse implementation.