Add GeneratorFactory base class
Removes the burden of exposing generator class variables from
factory implementations. Should the list or names of variables to
be exposed change in the future, this base class will be kept in
sync with such changes and so factories subclassing this base class
will not need to change their code.