Ensure that listener sees correct name when lexer rule refers to another lexer rule
Since token sub-rules (e.g., fragments) are not represented by
unique nodes in the tree, even though listeners are called when the
generation of a sub-rule starts or ends, the listeners only see the
top-level token node and the name of the top-level lexer rule.

This commit changes this by faking the top-level token node to have
the name of the sub-rule rule while generating the sub-rule.

Co-authored-by: Renata Hodovan <reni@inf.u-szeged.hu>