Limit the dot single-character wildcard to ASCII in the example fuzzer (#271)
Allowing any Unicode character makes the fuzzer depend on the
machine where the fuzzer is generated (i.e., on the version of its
Unicode library). As the development happens on multiple machines,
this can cause too much unnecessary churn in the code base.

Moreover, having long Unicode code point ranges in the example is
not helping the presentation purposes.