Add the grammarinator- prefix to the name of the generate-* binary
When building the C++ version, the binary executable tool that
generates test cases was named `generate-SUFFIX` (where `SUFFIX`
stands for the grammar/format used). To make the naming more
aligned with the Python version and to make the tool's relation to
Grammarinator clearer, the name of the tool is prefixed with
`grammarinator-`.

Also added directives on how to install the tool if the project is
installed.