Change C++ build helper to resolve install prefix relative to CWD instead of project root (#310)
The include directory is already turned into an absolute path
relative to the current working directory (when `build.py` is
called). However, the install prefix (if given) has been directly
passed to cmake -- but the CWD for cmake has been changed to the
root of the C++ code base, and so the install prefix was always
interpreted relative to the C++ project root.

This could cause ambiguity on the command line if `build.py` was
not executed from the root of the C++ code base. (Two directories
had to be specified relative to two different bases.)

This change helps unify this by also turning the install prefix
into an absolute path relative to CWD before passing it to cmake.