Run sphinx-build from the docs directory when toxing (#311)
This is to align the local current working directory of Sphinx with
the one used on ReadTheDocs. This ensures that runcmd directives
find programs to be executed consistently on the same paths. (And
so, we don't get "No such file or directory" error messages instead
of nice command outputs.)

(The change is somewhat speculative, but there is no way of testing
it other than pushing it to RTD and crossing fingers.)