Refactor Population and Individual implementations
The minimal requirements for an Individual is to provide or
load the root of its represented tree and be able to compute Annotations.
All other behavior should be implemented in subclasses as needed.

From now, the base Individual no longer expects a `name` parameter (used only by
FileIndividual), and instead allows directly providing and returning the
tree root. This makes the LibFuzzerIndividual class unnecessary, since
it required only this core functionality.