Do not require to define --generator in C++ build script
The generator can be defined either by the --generator cli flag of
build.py or through the optionally included config file (--include
argument). However, until now, it was mandatory to define the
generator with --generator even if it was also defined (or redefined)
by the include file as well. Now, it's optional which approach is used.