Do not require a generator when building the C++ decoder only (#374)
Decoder in C++ only requires a reference to the serializer and the type
of treeformat to be used. It doesn't need any information about the
generator, model, listener, or transformer used for generating the
tree to be decoded. The patch eliminates such artificial requirements
from the build system and config.